refactor(browser): remove tab election runtime
