--- packages/jazz-tools/src/runtime/connection-manager/connection-roles/connection-role.ts
@@ -1,11 +1,0 @@
-import type { DurabilityTier } from "../../client.js";
-import type { ConnectionManagerClientInput } from "../types.js";
-
-export interface BrowserConnectionRole {
-  onClientCreated(input: ConnectionManagerClientInput): void;
-  ensureReady(tier?: DurabilityTier): Promise<void>;
-  disconnect(): Promise<void>;
-  reconnect(): Promise<void>;
-  updateAuth(authJson: string, sessionClaims: Record<string, unknown>): void;
-  shutdown(): Promise<void> | void;
-}
