+++ packages/jazz-tools/src/runtime/default-runtime-source.ts
@@ -23,7 +23,7 @@ import { parseJwtPayload, resolveClientSessionSync } from "./client-session.js";
 import type { WasmSchema } from "../drivers/types.js";
 import { httpUrlToWs } from "./url.js";
 import { authorBytesForSubject, isUsableSubject } from "./author-id.js";
-import { createBrokerFingerprint } from "./connection-manager/browser-broker-utils.js";
+import { createBrowserWorkerFingerprint } from "./browser-worker-config.js";
 import { getRuntimeSchemaCacheKey } from "../drivers/schema-wire.js";
 
 const DEFAULT_WASM_LOG_LEVEL = "warn";
@@ -185,7 +185,7 @@ export class DefaultRuntimeSource extends RuntimeSource<DbConfig> {
         logLevel: config.logLevel,
         telemetryCollectorUrl: config.telemetryCollectorUrl,
       },
-      createBrokerFingerprint(config, dbName, getRuntimeSchemaCacheKey(schema)),
+      createBrowserWorkerFingerprint(config, dbName, getRuntimeSchemaCacheKey(schema)),
       { onAuthFailure, onAuthRestored, onFailure, onStorageReset, onStorageInvalidated },
     );
   }
